refactor: extract llms.txt onPostBuild hook into own file
